test(hermetic-env): move the regression to its own discovered file
This file's own header says it in capitals — NEW TESTS GO IN A FILE OF THEIR
OWN, NOT IN A SECTION HERE, and the inline sections are history rather than a
pattern to copy. #3011 added one inline anyway. Caught by @coderabbitai after
the merge; this is the follow-up.

The stated reason is collision cost: six contributors picking section `60a` in
Aug-2026 forced about fifteen rebases and six serialized CI runs for six lines
of test code. There is a second one worth recording — an inline check cannot be
run by `--only`, which this same header warns about, so the test could not be
exercised locally at all while it lived here. As a discovered file it runs under
`node test-all.mjs --only hermetic`.

hermeticGitEnv moves to tests/helpers.mjs rather than being exported from here,
because importing test-all.mjs runs the whole suite. Both existing fixtures keep
calling the one construction — a test that hand-rolled its own env would keep
passing if the pins were dropped, which is how the original gap got in, and that
property is worth more than the file it lives in. tests/ is already covered by
SYSTEM_PATHS as a directory, so the new file needs no registration.

No behaviour change: the two assertions are the merged ones, verbatim. Seen red
by deleting both `delete` lines from the helper — user.name reads back as
`ambient-leak` and the fixture's own `git config` write lands in the ambient file
(local=false ambient=true).
